requirements.txt
httpx>=0.28.1
python-dotenv>=1.0.0 
mcp[cli]>=1.2.0
asyncio>=3.4.3